requirements.txt
alembic==1.14.0
asttokens==3.0.0
blinker==1.9.0
click==8.1.7
decorator==5.1.1
executing==2.1.0
Faker==33.3.0
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
ipdb==0.13.13
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
minio==7.2.3
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dotenv==1.0.1
python-magic==0.4.27
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.36
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
